requirements.txt
absl-py==0.7.1
astor==0.8.0
atari-py==0.2.6
certifi==2019.6.16
cloudpickle==1.2.1
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
gym==0.13.1
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.1.0
protobuf==3.9.0
pyglet==1.3.2
PyYAML==5.1.1
scipy==1.3.0
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.32.2
Werkzeug==0.15.5
wincertstore==0.2
wrapt==1.11.2
